requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-tavily>=0.1.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=0.1.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
streamlit>=1.28.0 